+++ lib/common/io_bridge_object_store/src/lib.rs
@@ -9,8 +9,8 @@
 //! | [`BlobBackend`]  | Trait: a typed `ObjectStore` + a builder from a `Config`. |
 //! | `Arc<S>`         | The [`AsyncRead`] handle; any [`BlobBackend`] `S`.        |
 //! | [`BlobFile<A>`]  | Sync `UniversalRead` wrapper over an `AsyncRead` + path.  |
-//! | [`BridgeRuntime`]| Owns a Tokio runtime + a worker thread + request channel. |
-//! | `PipelineInner`  | Per-pipeline reply channel and slot bookkeeping.          |
+//! | [`BridgeRuntime`]| Cheap-to-clone owner of a Tokio runtime.                  |
+//! | `PipelineInner`  | Reply channel + slot-keyed owned destination buffers.     |
 //!
 //! Concrete backends live under [`backends`]: [`backends::aws`],
 //! [`backends::gcp`], and [`backends::azure`]. Each module declares its own
@@ -18,13 +18,16 @@
 //!
 //! # Data flow
 //!
-//! The caller (sync) lives on the left, the [`BridgeRuntime`] worker thread
-//! (async) lives on the right. The two MPSC channels run between them, with
-//! request going right and reply going left.
+//! The caller (sync) lives on the left, the Tokio runtime (async) on the right.
+//! There is no intermediary dispatcher thread or request channel: a single read
+//! is driven inline via [`Handle::block_on`](tokio::runtime::Handle::block_on),
+//! and a batched read is handed to the runtime with
+//! [`Handle::spawn`](tokio::runtime::Handle::spawn). Spawned tasks report their
+//! result over the pipeline's single reply channel.
 //!
 //! ```text
-//!     CALLER THREAD  (sync)                         WORKER THREAD  (async, Tokio)
-//!     ═════════════════════                         ═════════════════════════════
+//!     CALLER THREAD  (sync)                       RUNTIME WORKER THREADS  (async, Tokio)
+//!     ═════════════════════                       ═════════════════════════════════════
 //!
 //!   ┌────────────────────────────────┐
 //!   │ BlobFile<A>                    │
@@ -36,46 +39,40 @@
 //!                  ▼
 //!   ┌──────────────────────────────────────────┐
 //!   │ PipelineInner<T, U>                      │
-//!   │   tx, rx  : reply channel                │
-//!   │   pending : AHashMap<u64, (U, Vec<T>)>   │
-//!   │   next_slot : u64                        │
+//!   │   tx, rx : reply channel                 │
+//!   │   slots  : PendingSlots<T, U>            │
+//!   │           (slot -> (U, Vec<T>))          │
 //!   └──────────────────────────────────────────┘
 //!                  │ schedule(rt, user_data, Vec<T>, future)
 //!                  │  ─ allocates Vec<T> of exact size
-//!                  │  ─ moves it into `pending`
+//!                  │  ─ moves it into `slots`
 //!                  │  ─ derives a SendBytePtr captured by `future`
-//!                  ▼
-//!   ┌────────────────────────────────┐                    ┌────────────────────────────────┐
-//!   │ BridgeRequest {                │   request channel  │ worker_loop:                   │
-//!   │   future,                      │ ─────────────────► │   req = request_rx             │
-//!   │   tx: reply_tx.clone(),        │ rt.tx().try_send() │           .recv().await        │
-//!   │   slot,                        │                    │                                │
-//!   │ }                              │                    │   tokio::spawn(async move {    │
-//!   └────────────────────────────────┘                    │     // writes into the pending │
-//!                                                         │     // Vec<T> via SendBytePtr  │
-//!                                                         │     // + AlignedBufWriter      │
-//!                                                         │     status = req.future.await  │
-//!                                                         │     req.tx.send(               │
-//!   ┌────────────────────────────────┐                    │       BridgeResponse {         │
-//!   │ wait():                        │    reply channel   │         slot,                  │
-//!   │   response =                   │ ◄───────────────── │         result: status,        │
-//!   │     reply_rx.blocking_recv()   │  spawned task uses │       },                       │
-//!   │   (user_data, buf) =           │  req.tx (clone of  │     ).await                    │
-//!   │     pending.remove(&slot)      │  pipeline reply_tx)│   })                           │
-//!   │   response.result?             │                    └────────────────────────────────┘
+//!                  ▼                                       ┌────────────────────────────────┐
+//!   ┌────────────────────────────────┐  rt.handle()       │ spawned task:                  │
+//!   │ reply_tx = tx.clone()          │  .spawn(task)      │   // writes into the pending   │
+//!   │ slots.insert(user, buf) ->slot │ ─────────────────► │   // Vec<T> via SendBytePtr    │
+//!   └────────────────────────────────┘                    │   // + AlignedBufWriter        │
+//!                                                         │   status = future.await        │
+//!                                                         │   reply_tx.send(               │
+//!   ┌────────────────────────────────┐    reply channel   │     BridgeResponse{slot,status})│
+//!   │ wait():                        │ ◄───────────────── │   ).await                      │
+//!   │   response =                   │  spawned task uses └────────────────────────────────┘
+//!   │     reply_rx.blocking_recv()   │  reply_tx (clone of pipeline reply_tx)
+//!   │   (user_data, buf) =           │
+//!   │     slots.remove(&slot)        │
+//!   │   response.result?             │
 //!   │   return Some((user_data, buf))│
 //!   └────────────────────────────────┘
 //!
 //!   Single-read fast path: BlobFile::read calls runtime.block_on(inner.read_range(..))
-//!   directly on the Tokio handle — no pipeline, no slot, no worker hop.
+//!   directly on the calling thread — no pipeline, no slot, no spawn.
 //! ```
 //!
 //! # Design notes
 //!
-//! - **Two channels.** Each [`BridgeRuntime`] owns one request channel; each
-//!   pipeline owns one reply channel. A [`BridgeRequest`] carries the reply
-//!   sender as a return address, so the worker has no routing table — it
-//!   replies on the sender it was handed.
+//! - **One reply channel per pipeline.** Each pipeline owns a reply channel and
+//!   clones its sender into every spawned task, which uses it as a return
+//!   address — so there is no routing table.
 //!
 //! - **Runtime decoupling.** A pipeline does not store a [`BridgeRuntime`].
 //!   The runtime is supplied per `schedule` call, allowing one pipeline to
@@ -87,9 +84,8 @@
 //!   context.
 //!
 //! - **No `dyn` in the backend.** [`AsyncRead`] is implemented for `Arc<S>`,
-//!   not `Arc<dyn ObjectStore>`. The only remaining type erasure is the boxed
-//!   future in [`BridgeRequest::future`] — required because struct fields
-//!   cannot hold `impl Trait` and the channel must carry one concrete type.
+//!   not `Arc<dyn ObjectStore>`. Spawning takes the future by generic type, so
+//!   no boxing or type erasure is needed at the dispatch boundary either.
 //!
 //! - **No `Bytes` aggregation in the pipeline.** The pipeline owns the typed
 //!   destination `Vec<T>` for every in-flight request. The future writes the
@@ -114,4 +110,4 @@ pub use backend::BlobBackend;
 pub use file::BlobFile;
 pub use pipeline::{BorrowedBlobPipeline, OwnedBlobPipeline};
 pub use read::AsyncRead;
-pub use runtime::{BridgeRequest, BridgeResponse, BridgeRuntime};
+pub use runtime::BridgeRuntime;
