+++ lib/common/io_bridge_object_store/src/pipeline/mod.rs
@@ -1,0 +1,22 @@
+//! Read pipelines bridging the sync universal-IO pipeline traits to the async
+//! object-store backends.
+//!
+//! - [`slots`]: caller-side slot bookkeeping that owns the destination buffers
+//!   ([`PendingSlots`](slots::PendingSlots)).
+//! - [`buffer`]: the `Send` raw-pointer primitive and the shared read-future
+//!   builder that streams a backend read straight into a destination `Vec<T>`.
+//! - [`inner`]: the schedule/wait engine ([`PipelineInner`](inner::PipelineInner))
+//!   shared by both public pipelines.
+//! - [`borrowed`] / [`owned`]: the [`BorrowedBlobPipeline`] /
+//!   [`OwnedBlobPipeline`] trait impls over a [`BlobFile`](crate::BlobFile).
+
+mod borrowed;
+mod buffer;
+mod inner;
+mod owned;
+mod slots;
+
+pub use borrowed::BorrowedBlobPipeline;
+pub use owned::OwnedBlobPipeline;
+
+pub(crate) const BLOB_PIPELINE_CAPACITY: usize = 256;
