+++ lib/segment/src/common/operation_error.rs
@@ -165,8 +165,8 @@ impl From<UniversalIoError> for OperationError {
             | UniversalIoError::Uninitialized { .. }
             | UniversalIoError::QueueIsFull
             | UniversalIoError::S3(_)
-            | UniversalIoError::S3RuntimeShutDown
-            | UniversalIoError::S3Config { .. } => Self::service_error(err.to_string()),
+            | UniversalIoError::S3Config { .. }
+            | UniversalIoError::TaskPanicked(_) => Self::service_error(err.to_string()),
         }
     }
 }
